Added connection with PokeApi and delete button
Created a searchPokemon function that displays the name of a pokemon on the right display of the pokedex by requesting data from the PokeApi when pressing the button that has onclick = searchPokemon and based on the inputvalue that is changed to type = text. Also added the PokeApi link in the HTML file. To clear the result that is displayed with the searchPokemon button, I created a delete function on the green button.